requirements.txt
bcrypt==3.2.0
bidict==0.21.3
cffi==1.14.6
click==8.0.1
Flask==2.0.1
Flask-SocketIO==4.3.1
greenlet==1.1.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pkg-resources==0.0.0
psycopg2==2.9.1
pycparser==2.20
python-decouple==3.4
python-engineio==3.13.2
python-socketio==4.6.0
six==1.16.0
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
